Changes based on PR-37 feedback; factor out portal calls to portal_calls.py.
